Using a singleton for handling multiple configs and status tracking
