+++ src/app_genesis/routes/__init__.py
@@ -1,4 +1,5 @@
 from .browser import browser
+from .config_editing import config_handler
 from .runner import runner
 
-__all__ = ["browser", "runner"]
+__all__ = ["browser", "config_handler","runner"]
