+++ src/app_genesis/routes/runner.py
@@ -7,27 +7,26 @@ from flask import (
     render_template,
     request
 )
-from .genesis_runner import GenesisRunner
+
+from configs_registry import ConfigRegistry
 
 runner = Blueprint("runner", __name__)
 
 CONFIG_DIR = Path("configs").resolve()
 OUTPUT_DIR = Path("output").resolve()
+config_registry = ConfigRegistry()
 
-genesis_runner = GenesisRunner()
+#genesis_runner = GenesisRunner()
 
 # Centrale in-memory status store
 # voorbeeld: {"config1.yaml": "running", "config2.yaml": "finished"}
-running_configs = {}
 
 
 @runner.route("/run/<filename>")
 def config_run(filename: str) -> Response:
-    """Start een GenesisRunner-proces voor een opgegeven configuratiebestand.
-
-    Valideert het configuratiebestand, start het proces indien nodig en markeert de status als actief. Foutmeldingen worden weergegeven als het bestand ongeldig is of het proces niet kan worden gestart.
-    """
+    """Start een GenesisRunner-proces voor een opgegeven configuratiebestand."""
     config_file_path = (CONFIG_DIR / filename).resolve()
+
     # Veiligheidscheck: bestand moet binnen CONFIG_DIR liggen
     if not str(config_file_path).startswith(str(CONFIG_DIR)):
         error_message = "Ongeldig configuratiepad."
@@ -37,26 +36,26 @@ def config_run(filename: str) -> Response:
         error_message = f"Configuratiebestand '{filename}' niet gevonden."
         return render_template("error.html", message=error_message), 404
 
+    if running_configs.get(filename) == "running":
+        # Als het proces al draait, geef een foutmelding terug
+        error_message = f"Configuratiebestand '{filename}' draait al."
+        return render_template("error.html", message=error_message), 400
+
     running_configs[filename] = "running"  # Markeer als actief
 
-    if not genesis_runner.is_running():
-        try:
-            genesis_runner.start(path_config=config_file_path)
-        except Exception as e:
-            running_configs[filename] = "error"
-            error_message = f"Fout bij het starten van de runner: {str(e)}"
-            return render_template("error.html", message=error_message), 500
+    try:
+        genesis_runner.start(path_config=config_file_path)
+    except Exception as e:
+        running_configs[filename] = "error"
+        error_message = f"Fout bij het starten van de runner: {str(e)}"
+        return render_template("error.html", message=error_message), 500
 
-    return render_template("run.html", config=filename)
+    return render_template("runner.html", config=filename)
 
 
 @runner.route("/stream")
 def stream() -> Response:
-    """Streamt de uitvoer van de GenesisRunner naar de client als Server-Sent Events.
-
-    Deze functie zet de uitvoer van het GenesisRunner-proces om naar HTML en stuurt deze regel voor regel naar de client.
-    Speciale signalen worden verzonden als er om input wordt gevraagd of als het proces is afgerond.
-    """
+    """Streamt de uitvoer van de GenesisRunner naar de client als Server-Sent Events."""
     def generate():
         conv = Ansi2HTMLConverter(inline=True)
         try:
@@ -65,7 +64,8 @@ def stream() -> Response:
                 yield f"data: {html_line}\n\n"
 
                 if "doorgaan" in html_line or "antwoorden" in html_line:
-                    yield "data: asking_question\n\n"
+                    running_configs[cfg] = "waiting_input"
+                    yield f"data: waiting_input|{cfg}\n\n"
                 elif "Afgerond" in html_line:
                     yield "data: finished\n\n"
                     # Zoek actieve config en markeer als klaar
@@ -82,7 +82,14 @@ def stream() -> Response:
 
 
 @runner.route("/check_running_status", methods=["POST"])
-def check_running_status():
+def check_running_status() -> Response:
+    """Controleert of een opgegeven configuratie momenteel actief is.
+
+    Ontvangt de bestandsnaam van de client en retourneert of deze configuratie als 'running' is gemarkeerd.
+
+    Returns:
+        Response: Een JSON-object met de sleutel 'is_running' die aangeeft of de configuratie actief is.
+    """
     data = request.get_json()
     filename = data.get("filename")
     is_running = running_configs.get(filename) == "running"
@@ -90,13 +97,19 @@ def check_running_status():
 
 
 @runner.route("/status/all", methods=["GET"])
-def status_all():
-    """Geeft een overzicht van de status van alle bekende configuraties."""
+def status_all() -> Response:
+    """Geeft een overzicht van de status van alle bekende configuraties.
+
+    Retourneert een JSON-object met de status van alle momenteel bekende configuratiebestanden.
+
+    Returns:
+        Response: Een JSON-object met de status van alle configuraties.
+    """
     return jsonify(running_configs)
 
 
 @runner.route("/input", methods=["POST"])
-def send_input() -> Response:
+def input() -> Response:
     """Ontvangt invoer van de client en stuurt deze door naar de GenesisRunner."""
     value = request.json.get("value")
     if value is None or not isinstance(value, str) or not value.strip():
@@ -108,3 +121,22 @@ def send_input() -> Response:
         return jsonify({"status": "error", "message": f"Failed to send input: {str(e)}"}), 500
 
     return jsonify({"status": "ok"})
+
+
+@runner.route("/send_input", methods=["POST"])
+def send_input():
+    data = request.get_json()
+    filename = data.get("filename")
+    answer = data.get("answer", "")
+
+    if genesis_runner.process and genesis_runner.process.stdin:
+        try:
+            # stuur direct J of N door
+            genesis_runner.process.stdin.write((answer + "\n").encode("utf-8"))
+            genesis_runner.process.stdin.flush()
+            running_configs[filename] = "running"
+            return jsonify({"status": "ok"})
+        except Exception as e:
+            return jsonify({"status": "error", "message": str(e)}), 500
+    else:
+        return jsonify({"status": "error", "message": "Geen actief proces"}), 400
